# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: getTenantAdmin
      tags: [administration]
      summary: Get the current tenant administration scope and managed accounts
      responses:
        "200":
          description: Tenant-local administration snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantAdminSnapshot"
        "401":
          description: Authentication is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: The account has no administration scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "428":
          description: A permanent password must be set first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createManagedTenantAccount
      tags: [administration]
      summary: Create a module admin, module staff, or client account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagedTenantAccountCreateRequest"
      responses:
        "201":
          description: Managed account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedTenantAccount"
        "400":
          description: Invalid role or assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: The account cannot manage the requested role or module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: The username is already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/users/{accountId}:
    patch:
      operationId: updateManagedTenantAccountStatus
      tags: [administration]
      summary: Activate or deactivate a managed tenant account
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagedTenantAccountStatusRequest"
      responses:
        "200":
          description: Managed account status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedTenantAccountStatusResponse"
        "403":
          description: The account cannot manage the target account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: The managed account does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteManagedTenantAccount
      tags: [administration]
      summary: Permanently delete a module staff or client account and its assignments
      parameters:
        - name: accountId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Managed account and assignments deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffAccountDeletionResponse"
        "403":
          description: The exact module admin is required
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "404":
          description: The managed account does not exist
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  /admin/users/{accountId}/access:
    patch:
      operationId: updateManagedTenantAccountAccess
      tags: [administration]
      summary: Change a staff or client role and workspace access
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagedTenantAccountAccessRequest"
      responses:
        "200":
          description: Managed account access updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedTenantAccount"
        "400":
          description: Invalid role or workspace assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: The account cannot manage the target account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: The managed account does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/users/{accountId}/reset-password:
    post:
      operationId: resetManagedTenantAccountPassword
      tags: [administration]
      summary: Set a temporary password on an active managed tenant account
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagedTenantAccountPasswordResetRequest"
      responses:
        "200":
          description: Temporary password set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedTenantAccountPasswordResetResponse"
        "403":
          description: The account cannot manage the target account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: The active managed account does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /platform-staff/login:
    post:
      operationId: loginPlatformStaff
      tags: [administration]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/LoginRequest" } } } }
      responses:
        "200": { description: Authenticated platform staff account, content: { application/json: { schema: { $ref: "#/components/schemas/PlatformStaffSession" } } } }
        "400": { description: Invalid credentials payload, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "401": { description: Invalid platform staff credentials, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
...
  /platform-staff/me:
    get:
      operationId: getCurrentPlatformStaff
      tags: [administration]
      responses:
        "200": { description: Current platform staff account, content: { application/json: { schema: { $ref: "#/components/schemas/PlatformStaffSession" } } } }
        "401": { description: Platform staff authentication is required, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
  /platform-staff/change-password:
    post:
      operationId: changePlatformStaffPassword
      tags: [administration]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ChangePasswordRequest" } } } }
      responses:
        "200": { description: Password changed, content: { application/json: { schema: { $ref: "#/components/schemas/PlatformStaffPasswordChangeResponse" } } } }
        "400": { description: Invalid password change payload, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "401": { description: Authentication or current password is invalid, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
  /platform-staff/workspaces:
    get:
      operationId: getPlatformStaffWorkspaces
      tags: [administration]
      responses:
        "200": { description: Assigned active platform workspaces, content: { application/json: { schema: { $ref: "#/components/schemas/ManagedRegistryWorkspaceList" } } } }
        "401": { description: Platform staff authentication is required, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "403": { description: A password change is required, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
  /platform-staff/workspaces/{workspaceKey}:
    get:
      operationId: getPlatformStaffWorkspace
      tags: [administration]
      parameters: [{ name: workspaceKey, in: path, required: true, schema: { $ref: "#/components/schemas/PlatformWorkspaceKey" } }]
      responses:
        "200": { description: Assigned platform workspace, content: { application/json: { schema: { $ref: "#/components/schemas/ManagedRegistryWorkspace" } } } }
        "401": { description: Platform staff authentication is required, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "403": { description: A password change is required, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "404": { description: Workspace is not assigned or does not exist, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
...
  /owner/platform-staff/{platformStaffId}/workspaces:
    put:
      operationId: updatePlatformStaffWorkspaces
      tags: [administration]
      parameters: [{ name: platformStaffId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/PlatformStaffWorkspacesRequest" } } } }
      responses:
        "200": { description: Platform staff workspace assignments updated, content: { application/json: { schema: { $ref: "#/components/schemas/PlatformStaffWorkspacesResponse" } } } }
        "400": { description: Invalid staff identifier or workspace assignment, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "401": { description: Platform Administrator session is required, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "404": { description: Platform staff not found, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
...
components:
  schemas:
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 255
        password:
          type: string
          minLength: 1
          maxLength: 255
      required:
        - username
        - password
...
    ChangePasswordRequest:
      type: object
      additionalProperties: false
      properties:
        currentPassword:
          type: string
          minLength: 1
          maxLength: 255
        newPassword:
          type: string
          minLength: 8
          maxLength: 255
      required:
        - currentPassword
        - newPassword
...
    ModuleKey:
      type: string
      enum:
        - module_a
        - module_b
        - module_c
        - module_d
        - module_e
        - module_f
        - module_g
        - module_h
    WorkspaceKey:
      type: string
      pattern: "^ws-[1-9][0-9]*$"
...
    StaffAccountDeletionResponse:
      type: object
      properties:
        accountId: { type: string, format: uuid }
        deleted: { type: boolean, enum: [true] }
      required: [accountId, deleted]
...
    ManagedTenantAccount:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum:
            - module_admin
            - module_staff
            - client
        moduleKey:
          $ref: "#/components/schemas/ModuleKey"
        workspaceKeys:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceKey"
        isActive:
          type: boolean
        requiresPasswordChange:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - displayName
        - role
        - moduleKey
        - workspaceKeys
        - isActive
        - requiresPasswordChange
        - createdAt
    TenantAdminCurrentUser:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
        username:
          type: string
        role:
          type: string
          enum:
            - tenant_admin
            - module_admin
        moduleKey:
          oneOf:
            - $ref: "#/components/schemas/ModuleKey"
            - type: "null"
      required:
        - accountId
        - username
        - role
        - moduleKey
    TenantAdminSnapshot:
      type: object
      properties:
        tenantId:
          type: string
        subdomain:
          type: string
        customerName:
          type: string
        enabledModules:
          type: array
          items:
            $ref: "#/components/schemas/ModuleKey"
        currentUser:
          $ref: "#/components/schemas/TenantAdminCurrentUser"
        workspaces:
          type: array
          items:
            $ref: "#/components/schemas/ManagedWorkspaceOption"
        users:
          type: array
          items:
            $ref: "#/components/schemas/ManagedTenantAccount"
      required:
        - tenantId
        - subdomain
        - customerName
        - enabledModules
        - currentUser
        - workspaces
        - users
    ManagedWorkspaceOption:
      type: object
      properties:
        moduleKey:
          $ref: "#/components/schemas/ModuleKey"
        workspaceKey:
          $ref: "#/components/schemas/WorkspaceKey"
        displayName:
          type: string
      required:
        - moduleKey
        - workspaceKey
        - displayName
    ManagedTenantAccountCreateRequest:
      type: object
      additionalProperties: false
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 255
        displayName:
          type: string
          minLength: 1
          maxLength: 255
        role:
          type: string
          enum:
            - module_admin
            - module_staff
            - client
        moduleKey:
          $ref: "#/components/schemas/ModuleKey"
        workspaceKeys:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceKey"
        temporaryPassword:
          type: string
          minLength: 8
          maxLength: 255
      required:
        - username
        - displayName
        - role
        - moduleKey
        - workspaceKeys
        - temporaryPassword
    ManagedTenantAccountAccessRequest:
      type: object
      additionalProperties: false
      properties:
        role:
          type: string
          enum:
            - module_staff
            - client
        workspaceKeys:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/WorkspaceKey"
      required:
        - role
        - workspaceKeys
    WorkspaceAccessLevel:
      type: string
      enum:
        - active
        - sign_only
        - view_only
        - not_available
    WorkspaceContentNodeType:
      type: string
      enum:
        - page
        - tab
        - card
...
    WorkspaceContentNode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        parentId:
          oneOf:
            - type: string
              format: uuid
            - type: "null"
        semanticId:
          type: string
          pattern: "^(page|tab|card):[a-z0-9](?:[a-z0-9-]{0,126})$"
        parentSemanticId:
          oneOf:
            - type: string
              pattern: "^(page|tab|card):[a-z0-9](?:[a-z0-9-]{0,126})$"
            - type: "null"
        type:
          $ref: "#/components/schemas/WorkspaceContentNodeType"
        key:
          type: string
        displayName:
          type: string
        sortOrder:
          type: integer
        accessLevel:
          $ref: "#/components/schemas/WorkspaceAccessLevel"
      required:
        - id
        - parentId
        - semanticId
        - parentSemanticId
        - type
        - key
        - displayName
        - sortOrder
        - accessLevel
...
    ManagedTenantAccountStatusRequest:
      type: object
      additionalProperties: false
      properties:
        active:
          type: boolean
      required:
        - active
    ManagedTenantAccountStatusResponse:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
        active:
          type: boolean
      required:
        - accountId
        - active
    ManagedTenantAccountPasswordResetRequest:
      type: object
      additionalProperties: false
      properties:
        temporaryPassword:
          type: string
          minLength: 8
          maxLength: 255
      required:
        - temporaryPassword
    ManagedTenantAccountPasswordResetResponse:
      type: object
      properties:
        status:
          type: string
          enum:
            - password_reset
        accountId:
          type: string
          format: uuid
      required:
        - status
        - accountId
...
    WorkspaceType:
      type: string
      enum: [normal, public_information, contact_us]
...
    ManagedRegistryWorkspace:
      type: object
      properties:
        scope: { type: string, enum: [tenant, platform] }
        workspaceKey: { type: string }
        displayName: { type: string }
        workspaceType: { $ref: "#/components/schemas/WorkspaceType" }
        isActive: { type: boolean }
        publicVisible: { type: boolean }
        contactEnabled: { type: boolean }
        contentNodes:
          type: array
          items: { $ref: "#/components/schemas/WorkspaceContentNode" }
      required: [scope, workspaceKey, displayName, workspaceType, isActive, publicVisible, contactEnabled, contentNodes]
    ManagedRegistryWorkspaceList:
      type: object
      properties:
        workspaces:
          type: array
          items: { $ref: "#/components/schemas/ManagedRegistryWorkspace" }
      required: [workspaces]
...
    PlatformWorkspaceKey:
      type: string
      pattern: "^pws-[1-9][0-9]*$"
    PlatformStaffSession:
      type: object
      properties:
        authenticated: { type: boolean, const: true }
        accountId: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        requiresPasswordChange: { type: boolean }
      required: [authenticated, accountId, username, displayName, requiresPasswordChange]
...
    PlatformStaffPasswordChangeResponse:
      type: object
      properties:
        status: { type: string, const: password_changed }
        requiresPasswordChange: { type: boolean, const: false }
      required: [status, requiresPasswordChange]
    PlatformStaffAccount:
      type: object
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        isActive: { type: boolean }
        requiresPasswordChange: { type: boolean }
        createdAt: { type: string, format: date-time }
        workspaceKeys:
          type: array
          items: { $ref: "#/components/schemas/PlatformWorkspaceKey" }
      required: [id, username, displayName, isActive, requiresPasswordChange, createdAt, workspaceKeys]
...
    PlatformStaffWorkspacesRequest:
      type: object
      additionalProperties: false
      properties:
        workspaceKeys: { type: array, minItems: 1, uniqueItems: true, items: { $ref: "#/components/schemas/PlatformWorkspaceKey" } }
      required: [workspaceKeys]
...
    PlatformStaffWorkspacesResponse:
      type: object
      properties:
        status: { type: string, const: platform_staff_workspaces_updated }
        platformStaff: { $ref: "#/components/schemas/PlatformStaffAccount" }
      required: [status, platformStaff]
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error